third_party/blink/web_tests/external/wpt/css/css-scroll-anchoring/scroll-padding-affects-anchoring-expected.txt

This is a testharness.js-based test.
FAIL scroll anchoring accounts for scroll-padding assert_equals: Shouldn't anchor to #changer, since it's covered by scroll-padding expected 150 but got 50
Harness: the test ran to completion.
